+++ playback/player/qt/extension/qmldir
@@ -1,0 +1,2 @@
+module extension
+plugin qmlplayerextension
