+++ playback/player/qt/extension/quickrenderer.cpp
