+++ playback/player/qt/qt.pro
@@ -1,0 +1,5 @@
+TEMPLATE = subdirs
+SUBDIRS = extension play
+
+play.depends = extension
+play.file = play.pro
